change migration

+++ migrations/sqls/20220726152644-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
   id SERIAL PRIMARY KEY,
   name VARCHAR(200),
-  price INTEGER,
+  price DECIMAL(10,2),
   category VARCHAR(200)
 );
